+++ src/components/article/ShareBt.js
@@ -1,24 +1,46 @@
 'use strict'
-import { FacebookButton, TwitterButton } from 'react-social'
-import classNames from 'classnames'
-import styles from './ShareBt.scss'
+import MobileDetect from 'mobile-detect'
 import React from 'react' // eslint-disable-next-line
+import styles from './ShareBt.scss'
+import { FacebookButton, TwitterButton } from 'react-social'
+
+class ShareBt extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      isDesktop: true
+    }
+  }
+
+  componentDidMount() {
+    let md = new MobileDetect(window.navigator.userAgent)
+    this.setState({
+      isDesktop: (!md.mobile() && !md.tablet())
+    })
+  }
 
 
-export const ShareBt = (props) => {
-  const { appId, fbIcon, lineIcon, url, title, twitterIcon } = props
-  const lineUrl = `http://line.naver.jp/R/msg/text/?${encodeURI(title + ' ' + url)}`
-  return (
-    <div className={styles['share-bt-container']}>
-      <FacebookButton className={styles.bt} url={url} appId={appId}>
-        <img src={fbIcon} />
-      </FacebookButton>
-      <TwitterButton className={styles.bt} message={title} url={url}>
-        <img src={twitterIcon} />
-      </TwitterButton>
-      <a href={lineUrl} className={classNames(styles.bt, styles.line)}>
+  render() {
+    const { isDesktop } = this.state
+    const { appId, fbIcon, lineIcon, url, title, twitterIcon } = this.props
+    const lineUrl = `http://line.naver.jp/R/msg/text/?${encodeURI(title + ' ' + url)}`
+    const lineJsx = isDesktop ? null : (
+      <a href={lineUrl} className={styles.bt}>
         <img src={lineIcon} />
       </a>
-    </div>
-  )
+    )
+    return (
+      <div className={styles['share-bt-container']}>
+        <FacebookButton className={styles.bt} url={url} appId={appId}>
+          <img src={fbIcon} />
+        </FacebookButton>
+        <TwitterButton className={styles.bt} message={title} url={url}>
+          <img src={twitterIcon} />
+        </TwitterButton>
+        {lineJsx}
+      </div>
+    )
+  }
 }
+
+export { ShareBt }