+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-### Unreleased 
+### Unreleased
 
-### 1.1.3
+### 1.1.4
+- [Update] Update src/utils/image-processor.js. Add handling for empty or wrong parameter.
+- [Update] Given empty string as a fallback for article title.
+- [Update] Decrease redis expire time from 1 hour to 5 mins.
+- [Feature] Add print icon for friendly printing.
+
+### 1.1.3 
 - [Update] Change link of the animal protection project
 - [Bug] Top description cannot be displayed properly if the content brief has been edited